+++ RELEASE_NOTES.md
@@ -3,15 +3,33 @@
 ## Version TBD
 _Date TBD_
 
+### Features Added
+* Changed certification status names previously known as Suspended, Terminated, and Withdrawn.
+
+### Bugs Fixed
+
+---
+
+## Version 5.0.0
+_19 September 2016_
+
 ### Features added
 * Add optional argument 'edition' to /download call.
   * Generate the chpl-all file as well as a chpl-{edition} file for each edition present in the database for download.
   * Downloads chpl-all by default
 * Allowed c3/c4 to be connected to CQMs
+* Added error messages for c1/2/3/4-CQM mismatch
+* Re-factored Certification Id web services to improve performance
 * Added statuses object that shows aggregate number of certified products associated with each developer and product. These objects will allow the website search page to filter on a developer/product's number of certified products that are active/retired/withdrawn/suspended/terminated.
+* Updated /developers call to be faster
+* Changed /developers/update to update transparency attestation based on acb name, not acb ID
+* Re-factored Certification Id API controller to support new operations ("create", "search", "verify(GET)")
+* Changed /certification_ids/all operation to just /certification_ids
 
 ### Bugs fixed
 * Pending 2015 products can now have CQM versions modified
+* API Documentation Link is now correctly validated on upload and edit 
+* Added new style product number to CAP Activity Report descriptions
 
 ---
 
@@ -25,7 +43,10 @@ _30 August 2016_
 * TO DO DURING THE RELEASE: change the questionable activity email recipients to just be the ONC_CHPL@hhs.gov email (i.e. remove onc.certification@hhs.gov)
 * TO DO DURING THE RELEASE: create a file (can call it cleantomcat) in /etc/cron.daily and chmod a+x the file. Contents of the file are listed as a comment in OCD-811. The command deletes files that have not been written to since X days ago. The number near the end of the command is X.
 * Updated /activity API endpoint to incorporate new parameters to filter by API-Key, sort dateAscending, and filter by start & end date.
+<<<<<<< HEAD
 
+=======
+>>>>>>> upstream/development
 * Do not allow 170.315 (d)(3) to mark GAP as true
 * Added 'responsibleUser' field with all user data for /activity reports
 * Removed CORSFilter in web.xml; this was preventing some ajax calls from other domains